Assignment_2/requirements.txt
torch
torchvision
numpy
PyMCubes
matplotlib